add quantity badge component and interface
