+++ react/QuantityBadge.tsx
@@ -1,0 +1,3 @@
+import QuantityBadge from './components/QuantityBadge'
+
+export default QuantityBadge;
