+++ react/components/MinicartIconButton.tsx
@@ -1,10 +1,9 @@
 import React from 'react'
 import { ButtonWithIcon } from 'vtex.styleguide'
-import { useOrderForm } from 'vtex.order-manager/OrderForm'
 
-import styles from '../styles.css'
 import { useMinicartCssHandles } from './CssHandlesContext'
 import { useMinicartDispatch, useMinicartState } from '../MinicartContext'
+import QuantityBadge from './QuantityBadge'
 
 export const CSS_HANDLES = [
   'minicartIconContainer',
@@ -17,49 +16,11 @@ interface Props {
   itemCountMode: MinicartTotalItemsType
 }
 
-const countCartItems = (
-  countMode: MinicartTotalItemsType,
-  allItems: OrderFormItem[]
-) => {
-  // Filter only main products, remove assembly items from the count
-  const items = allItems.filter(item => item.parentItemIndex === null)
-
-  if (countMode === 'distinctAvailable') {
-    return items.reduce((itemQuantity: number, item) => {
-      if (item.availability === 'available') {
-        return itemQuantity + 1
-      }
-      return itemQuantity
-    }, 0)
-  }
-
-  if (countMode === 'totalAvailable') {
-    return items.reduce((itemQuantity: number, item) => {
-      if (item.availability === 'available') {
-        return itemQuantity + item.quantity
-      }
-      return itemQuantity
-    }, 0)
-  }
-
-  if (countMode === 'total') {
-    return items.reduce((itemQuantity: number, item) => {
-      return itemQuantity + item.quantity
-    }, 0)
-  }
-
-  // countMode === 'distinct'
-  return items.length
-}
-
 const MinicartIconButton: React.FC<Props> = props => {
   const { Icon, itemCountMode, quantityDisplay } = props
-  const { orderForm, loading }: OrderFormContext = useOrderForm()
   const { handles } = useMinicartCssHandles()
   const { open, openBehavior, openOnHoverProp } = useMinicartState()
   const dispatch = useMinicartDispatch()
-  const quantity = countCartItems(itemCountMode, orderForm.items)
-  const itemQuantity = loading ? 0 : quantity
 
   const handleClick = () => {
     if (openOnHoverProp) {
@@ -84,23 +45,14 @@ const MinicartIconButton: React.FC<Props> = props => {
     dispatch({ type: open ? 'CLOSE_MINICART' : 'OPEN_MINICART' })
   }
 
-  const showQuantityBadge =
-    (itemQuantity > 0 && quantityDisplay === 'not-empty') ||
-    quantityDisplay === 'always'
+  
 
   return (
     <ButtonWithIcon
       icon={
         <span className={`${handles.minicartIconContainer} gray relative`}>
           <Icon />
-          {showQuantityBadge && (
-            <span
-              style={{ userSelect: 'none' }}
-              className={`${handles.minicartQuantityBadge} ${styles.minicartQuantityBadgeDefault} c-on-emphasis absolute t-mini bg-emphasis br4 w1 h1 pa1 flex justify-center items-center lh-solid`}
-            >
-              {itemQuantity}
-            </span>
-          )}
+          <QuantityBadge itemCountMode={itemCountMode} quantityDisplay={quantityDisplay}/>
         </span>
       }
       variation="tertiary"
