+++ store/interfaces.json
@@ -43,5 +43,8 @@
   },
   "minicart-close-button": {
     "component": "CloseButton"
+  },
+  "minicart-quantity-badge": {
+    "component": "QuantityBadge"
   }
 }
